Load CMIP5 data into climate4R grid format

Load CMIP5 data into climate4R grid format

Usage

loadCMIP5(
  indir,
  rcp,
  mod,
  vars = NULL,
  cmip5_vars = NULL,
  lon_lim = c(-11, 12),
  lat_lim = c(28, 44),
  years = 1991:2100,
  dictionary = system.file("extdata", "CMIP5_dictionary.csv", package = "dsclim")
)

Arguments

indir

A character string with the directory path for CMIP5 data. This directory should have a folder with the name of the RCP that is going to be loaded. The RCP folder should have in turn a folder with the name of the General Circulation Model that is goint to be loaded.

rcp

A character vector with the name of the RCP that is going to be loaded. Possible values are c("historical", "rcp2.6", "rcp4.5", "rcp6.0", "rcp8.5")

mod

A character vector with the name of the GCM that is going to be loaded. Possible values are c("CESM1-CAM5", "CSIRO-Mk3-6-0", "IPSL-CM5A-MR").

vars

A character vector with names of the CMIP5 variables in their standardized format (as they should match those from other datasets, e.g. TraCE21ka and UERRA datasets).

cmip5_vars

A character vector with names of the CMIP5 variables in their original format (as coded in original CMIP5 datasets).

lon_lim

A numeric vector (length = 2) with the longitudinal extent that is going to be loaded.

lat_lim

A numeric vector (length = 2) with the latitudinal extent that is going to be loaded.

years

A numeric vector with the sequence of years that are going to be loaded.

dictionary

A data dictionary (as in loadGridData) that is going to be used to load the data. By default it uses an internal dictionary that standardize CMIP5, TraCE21ka, and UERRA data to the same format.

Value

A list object with the structure suitable for the climate4R suite. More specifically the data are as returned from the loadGridData function.

Examples

if (FALSE) { # \dontrun{
dsclim::loadCMIP5("../../Data/CMIP5/",
  "rcp2.6",
  "CESM1-CAM5",
  lon_lim = c(-11, 12),
  lat_lim = c(28, 44),
  years = 2006:2100
)
} # }